project-path-apis/workers/requirements.txt
transformers==4.36.0
numpy
pdfplumber
Pillow
Werkzeug
celery
python-dotenv
redis
PyPDF2
python-multipart
openai
boto3
sentence-transformers
pgvector
SQLAlchemy
SQLAlchemy-Utils
Flask-SQLAlchemy
psycopg2-binary

